requirements.txt
# Testing and code checking
hypothesis
pytest
python-ta

# General packages
tqdm
colorama
texttable
hurry.filesize

# Data scraping
internetarchive

# Data processing and machine learning
contractions>0.0.18
tensorflow-text
tensorflow
tf-models-official
sklearn

# For string searching
suffix_trees

# Visualisation
dash==1.18.0
dash-bootstrap-components
matplotlib
tikzplotlib  # For exporting figures to latex

# K-hop graph algorithms
networkx  # Graph visualisation library
python-louvain  # Community detection on graphs
bezier==0.9.0  # Bezier curved edges for networkx
fa2  # ForceAtlas2 layout algorithm